+++ userspace-vdp/vdp-console8/cursor.h
@@ -2,23 +2,24 @@
 #define CURSOR_H
 
 #include <fabgl.h>
+#include <Arduino.h>
 
 #include "agon_keyboard.h"
 #include "vdp_protocol.h"
 #include "graphics.h"
 #include "viewport.h"
 
-extern uint8_t	fontW;
-extern uint8_t	fontH;
+extern int		fontW;
+extern int		fontH;
 extern void drawCursor(Point p);
-extern void scrollRegion(Rect * r, uint8_t direction, int16_t amount);
+extern void scrollRegion(Rect * r, int direction, int amount);
 
 Point			textCursor;						// Text cursor
 Point *			activeCursor;					// Pointer to the active text cursor (textCursor or p1)
 bool			cursorEnabled = true;			// Cursor visibility
-uint8_t			cursorBehaviour = 0;			// Cursor behavior
+byte			cursorBehaviour = 0;			// Cursor behavior
 bool 			pagedMode = false;				// Is output paged or not? Set by VDU 14 and 15
-uint8_t			pagedModeCount = 0;				// Scroll counter for paged mode
+int				pagedModeCount = 0;				// Scroll counter for paged mode
 
 
 // Render a cursor at the current screen position
@@ -41,7 +42,7 @@ inline void setActiveCursor(Point * cursor) {
 	activeCursor = cursor;
 }
 
-inline void setCursorBehaviour(uint8_t setting, uint8_t mask = 0xFF) {
+inline void setCursorBehaviour(byte setting, byte mask = 0xFF) {
 	cursorBehaviour = (cursorBehaviour & mask) ^ setting;
 }
 
@@ -71,12 +72,12 @@ void cursorDown() {
 		pagedModeCount++;
 		if (pagedModeCount >= (activeViewport->Y2 - activeViewport->Y1 + 1) / fontH) {
 			pagedModeCount = 0;
-			uint8_t ascii;
-			uint8_t vk;
-			uint8_t down;
+			byte ascii;
+			byte vk;
+			byte down;
 			if (!wait_shiftkey(&ascii, &vk, &down)) {
 				// ESC pressed
-				uint8_t packet[] = {
+				byte packet[] = {
 					ascii,
 					0,
 					vk,
@@ -139,7 +140,7 @@ void cursorHome() {
 
 // TAB(x,y)
 //
-void cursorTab(uint8_t x, uint8_t y) {
+void cursorTab(int x, int y) {
 	activeCursor->X = x * fontW;
 	activeCursor->Y = y * fontH;
 }
